Step 2: move cconstants.py co a common tests folder
