+++ tests/constants.py
@@ -6,16 +6,16 @@ Released under the terms of DataRobot Tool and Utility Agreement.
 """
 import os
 
-REPO_ROOT_PATH = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", ".."))
+REPO_ROOT_PATH = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 
 TESTS_ROOT_PATH = os.path.join(REPO_ROOT_PATH, "tests")
 TESTS_FIXTURES_PATH = os.path.join(TESTS_ROOT_PATH, "fixtures")
 TESTS_ARTIFACTS_PATH = os.path.join(TESTS_FIXTURES_PATH, "drop_in_model_artifacts")
 TESTS_DATA_PATH = os.path.join(TESTS_ROOT_PATH, "testdata")
 TESTS_DEPLOYMENT_CONFIG_PATH = os.path.join(TESTS_DATA_PATH, "deployment_config")
-TRANSFORM_TEMPLATES_PATH = os.path.join(TESTS_ROOT_PATH, "..", "task_templates", "1_transforms")
-ESTIMATORS_TEMPLATES_PATH = os.path.join(TESTS_ROOT_PATH, "..", "task_templates", "2_estimators")
-TRAINING_TEMPLATES_PATH = os.path.join(TESTS_ROOT_PATH, "..", "task_templates", "3_pipelines")
+TRANSFORM_TEMPLATES_PATH = os.path.join(REPO_ROOT_PATH, "task_templates", "1_transforms")
+ESTIMATORS_TEMPLATES_PATH = os.path.join(REPO_ROOT_PATH, "task_templates", "2_estimators")
+TRAINING_TEMPLATES_PATH = os.path.join(REPO_ROOT_PATH, "task_templates", "3_pipelines")
 
 MODEL_TEMPLATES_PATH = os.path.join(REPO_ROOT_PATH, "model_templates")
 PUBLIC_DROPIN_ENVS_PATH = os.path.join(REPO_ROOT_PATH, "public_dropin_environments")
