+++ tests/drum/test_dr_api_access.py
@@ -19,8 +19,8 @@ from datarobot_drum.drum.enum import ArgumentsOptions
 
 from datarobot_drum.resource.utils import _create_custom_model_dir
 from datarobot_drum.resource.utils import _exec_shell_cmd
-from tests.drum.constants import PYTHON_UNSTRUCTURED_DR_API_ACCESS
-from tests.drum.constants import UNSTRUCTURED
+from tests.constants import PYTHON_UNSTRUCTURED_DR_API_ACCESS
+from tests.constants import UNSTRUCTURED
 from tests.drum.utils import SimpleCache
