+++ tests/drum/test_fit.py
@@ -4,7 +4,6 @@ All rights reserved.
 This is proprietary source code of DataRobot, Inc. and its affiliates.
 Released under the terms of DataRobot Tool and Utility Agreement.
 """
-import json
 import os
 import shutil
 from tempfile import NamedTemporaryFile
@@ -22,7 +21,7 @@ from datarobot_drum.resource.utils import (
     _create_custom_model_dir,
     _exec_shell_cmd,
 )
-from .constants import (
+from tests.constants import (
     ANOMALY,
     BINARY,
     BINARY_INT,
