+++ tests/drum/test_fit_per_framework.py
@@ -26,7 +26,7 @@ from datarobot_drum.resource.utils import (
     _create_custom_model_dir,
     _exec_shell_cmd,
 )
-from .constants import (
+from tests.constants import (
     ANOMALY,
     BINARY,
     BINARY_BOOL,
