+++ tests/drum/test_performance_check.py
@@ -22,7 +22,7 @@ from datarobot_drum.resource.utils import (
 )
 
 
-from .constants import (
+from tests.constants import (
     SKLEARN,
     REGRESSION,
     BINARY,
