+++ tests/drum/test_stats.py
@@ -10,9 +10,7 @@ import pandas as pd
 import pytest
 import requests
 
-from datarobot_drum.drum.enum import ModelInfoKeys, TargetType
-from datarobot_drum.drum.description import version as drum_version
-from .constants import (
+from tests.constants import (
     PYTHON,
     REGRESSION,
     RESPONSE_PREDICTIONS_KEY,
